Refactored winapi type namespaces to align with source header files
